Add waiting animation function to HttpDialogProcessor
Also make it possible to use HttpDialogProcessor without creating your own subclass.
